wip scoring

+++ Ported/LabRat/Assets/Scripts/Systems/PlayerInitializationSystem.cs
@@ -64,21 +64,26 @@ public class PlayerInitializationSystem : SystemBase
                 All = new[] {ComponentType.ReadOnly<Player>()}
             });
             var cc = q.CalculateEntityCount();
-            var localP = Players = new NativeArray<Entity>(cc, Allocator.Persistent);
-            ecb = new EntityCommandBuffer(Allocator.Temp);
-            Entities.WithName("InitPlayerArray")
-                .WithNone<PlayerInitializedTag>()
-                .ForEach((int entityInQueryIndex, Entity e, in Player p) =>
-                {
-                    localP[entityInQueryIndex] = e;
-                    ecb.AddComponent<PlayerInitializedTag>(e);
-                }).Run();
-            ecb.Playback(EntityManager);
-            ecb.Dispose();
+            if (cc > 0)
+            {
+                var localP = Players = new NativeArray<Entity>(cc, Allocator.Persistent);
+                ecb = new EntityCommandBuffer(Allocator.Temp);
+                Entities.WithName("InitPlayerArray")
+                    .WithNone<PlayerInitializedTag>()
+                    .ForEach((int entityInQueryIndex, Entity e, in Player p) =>
+                    {
+                        localP[entityInQueryIndex] = e;
+                        ecb.AddComponent<PlayerInitializedTag>(e);
+                    }).Run();
+                ecb.Playback(EntityManager);
+                ecb.Dispose();
+
+                // Quick'n'dirty cleanup
+                if (EntityManager.HasComponent<GameStateCleanup>(GetSingletonEntity<PlayerInitialization>()))
+                    EntityManager.DestroyEntity(m_AnyPlayer);
+            }
         }
     
-        // Quick'n'dirty cleanup
-        if (EntityManager.HasComponent<GameStateCleanup>(GetSingletonEntity<PlayerInitialization>()))
-            EntityManager.DestroyEntity(m_AnyPlayer);
+  
     }
 }
